Finilize `Persistence` layer refactoring

+++ src/server/ContactSystem.Infrastructure.Persistence/Pagination/Interfaces/IPagedList.cs
@@ -2,13 +2,13 @@ namespace ContactSystem.Infrastructure.Persistence.Pagination.Interfaces;
 
 public interface IPagedList : IEnumerable
 {
-	int CurrentOffset { get; }
+	ulong CurrentOffset { get; }
 
-	int TotalPages { get; }
+	ulong TotalPages { get; }
 
-	int Limit { get; }
+	ulong Limit { get; }
 
-	int TotalCount { get; }
+	ulong TotalCount { get; }
 }
 
 public interface IPagedList<out T> : IPagedList, IEnumerable<T>
